requirements.txt
Flask==2.0.1
gunicorn==20.1.0
requests
wikipedia
line-bot-sdk
pandas
googletrans
schedule==1.1.0